+++ tests/conftest.py
@@ -20,6 +20,9 @@ os.environ["AIRTHINGS_SIMULATOR"] = "1"
 os.environ["AQICN_SIMULATOR"] = "1"
 os.environ["TEMPERATURE_BOT_DATABASE_ROOT"] = str(Path(__file__).parent)
 os.environ["DB_PATH"] = str(Path(__file__).parent / "temperature-bot.db")
+os.environ["TEMPERATURE_BOT_CONFIG"] = str(
+    Path(__file__).parent / "temperature-bot-config-test.yaml"
+)
 
 # pylint: disable=wrong-import-position
 from app import hubitat, routes_api